Move InstructorPage to actual-pages
