+++ src/actual-pages/InstructorsPage/index.ts
@@ -1,0 +1,1 @@
+export { InstructorsPage } from "./InstructorsPage";
